db 연동 2차 완료 (데이터 불러오기까지 가능)
