+++ app/src/main/java/com/example/royalroadofspec/community_detail.java
@@ -29,13 +29,17 @@ public class community_detail extends AppCompatActivity {
         Intent intent = getIntent();
 
         TextView title = (TextView) findViewById(R.id.detail_title);
+        TextView content = (TextView) findViewById(R.id.detail_content);
         TextView name = (TextView) findViewById(R.id.detail_name);
-        TextView board = (TextView) findViewById(R.id.detail_board);
+        TextView category = (TextView) findViewById(R.id.detail_category);
+        TextView date = (TextView) findViewById(R.id.detail_date);
 
 
         title.setText(intent.getStringExtra("title"));
-        name.setText(intent.getStringExtra("name"));
-        board.setText(intent.getStringExtra("board"));
+        content.setText(intent.getStringExtra("content"));
+        name.setText(intent.getStringExtra("userName"));
+        category.setText(intent.getStringExtra("category"));
+        date.setText(intent.getStringExtra("date"));
 
 
         LikeButton.setOnClickListener(new View.OnClickListener() {
