--- app/src/main/res/layout/activity_boards.xml
@@ -1,88 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".BoardsActivity">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-
-        <TextView
-            android:id="@+id/textView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="32dp"
-            android:layout_marginBottom="32dp"
-            android:text="파이어베이스 연동"
-            android:textSize="24dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.713" />
-
-        <EditText
-            android:id="@+id/et_board_title"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:layout_marginTop="300dp"
-            android:layout_marginBottom="20dp"
-            android:hint="제목을 입력하세요"
-            app:layout_constraintBottom_toTopOf="@id/et_board_content"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="16dp" />
-
-        <EditText
-            android:id="@+id/et_board_content"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:layout_marginTop="20dp"
-            android:layout_marginBottom="20dp"
-            android:hint="내용을 입력하세요"
-            app:layout_constraintBottom_toTopOf="@+id/et_board_userName"
-            app:layout_constraintTop_toBottomOf="@+id/et_board_title"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="16dp" />
-
-        <EditText
-            android:id="@+id/et_board_userName"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:layout_marginTop="100dp"
-            android:layout_marginBottom="302dp"
-            android:hint="닉네임을 입력하세요"
-            app:layout_constraintBottom_toTopOf="@+id/textView"
-            app:layout_constraintTop_toBottomOf="@id/et_board_content"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="16dp" />
-
-        <Button
-            android:id="@+id/btn_save"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="20dp"
-            android:text="저장"
-            app:layout_constraintBottom_toBottomOf="parent"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="0dp" />
-
-        <ListView
-            android:id="@+id/testList"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"></ListView>
-
-
-    </LinearLayout>
-
-</LinearLayout>
