+++ app/src/main/res/layout/activity_community_space4.xml
@@ -11,12 +11,12 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        tools:ignore="MissingConstraints"
-        tools:context=".community_space_4">
+        tools:context=".community_space_4"
+        tools:ignore="MissingConstraints">
 
         <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="861dp"
+            android:layout_height="match_parent"
             android:orientation="vertical">
 
             <TabWidget
@@ -29,16 +29,19 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content">
 
+
                 <LinearLayout
                     android:id="@+id/content1"
                     android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
+                    android:layout_height="match_parent"
                     android:orientation="vertical">
+
 
                     <ListView
                         android:id="@+id/infoList"
                         android:layout_width="match_parent"
                         android:layout_height="match_parent"></ListView>
+
 
                 </LinearLayout>
 
@@ -49,10 +52,12 @@
                     android:layout_height="match_parent"
                     android:orientation="vertical">
 
+
                     <ListView
                         android:id="@+id/questionList"
                         android:layout_width="match_parent"
                         android:layout_height="match_parent" />
+
 
                 </LinearLayout>
 
@@ -62,18 +67,6 @@
 
         </LinearLayout>
     </TabHost>
-
-    <Button
-        android:id="@+id/newActivity"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="다음페이지"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.95"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.976" />
 
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
